Use a vector of Posit Connect vanity URLs as a board

board_connect_url() lets you build up a board from individual vanity urls.

board_connect_url() is read only, and does not support versioning.

Usage

board_connect_url(
  vanity_urls,
  cache = NULL,
  use_cache_on_failure = is_interactive(),
  headers = connect_auth_headers()
)

connect_auth_headers(key = Sys.getenv("CONNECT_API_KEY"))

Arguments

vanity_urls

A named character vector of Connect vanity URLs. This board is read only, and the best way to write to a pin on Connect is board_connect().

cache

Cache path. Every board requires a local cache to avoid downloading files multiple times. The default stores in a standard cache location for your operating system, but you can override if needed.

use_cache_on_failure

If the pin fails to download, is it ok to use the last cached version? Defaults to is_interactive() so you'll be robust to poor internet connectivity when exploring interactively, but you'll get clear errors when the code is deployed.

headers

Named character vector for additional HTTP headers (such as for authentication). See connect_auth_headers() for Posit Connect support.

key

The Posit Connect API key.

Details

This board is a thin wrapper around board_url() which uses connect_auth_headers() for authentication via environment variable.

See also

Examples

if (FALSE) { # interactive()
connect_auth_headers()

board <- board_connect_url(c(
    my_vanity_url_pin = "https://colorado.posit.co/rsc/great-numbers/"
))

board %>% pin_read("my_vanity_url_pin")
}