Register RStudio Connect Board

Wrapper with explicit parameters over board_register() to register RStudio Connecet as a board.

board_register_rsconnect(name = "rsconnect", server = NULL,
  account = NULL, key = NULL, output_files = FALSE,
  cache = board_cache_path())

Arguments

name
Optional name for this board, defaults to 'rsconnect'.

server
Optional address to RStudio Connect server.

account
Optional account name to use with RStudio Connect.

key
The RStudio Connect API key.

output_files
Should the output in an automated report create output files?

cache
The local folder to use as a cache, defaults to board_cache_path().

See also

Examples

# NOT RUN {
# the following examples require an RStudio Connect API key

# register from rstudio
board_register_rsconnect()

# register from rstudio with multiple servers
board_register_rsconnect(server = "https://rstudio-connect-server")

# register from rstudio with multiple account
board_register_rsconnect(account = "account-name")

# register automated report for rstudio connect
board_register_rsconnect(key = Sys.getenv("RSTUDIO_KEY"),
                         server = "https://rstudio-connect-server")
# }