Register S3 Board

Wrapper with explicit parameters over board_register() to register an Amazon S3 bucket as a board.

board_register_s3(name = "s3", bucket = Sys.getenv("AWS_BUCKET"),
  key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
  cache = board_cache_path(), host = "s3.amazonaws.com", ...)

Arguments

name
Optional name for this board, defaults to 's3'.

bucket
The name of the Amazon S3 bucket. Defaults to the AWS_BUCKET environment variable.

key
The key of the Amazon S3 bucket. Defaults to the AWS_ACCESS_KEY_ID environment variable.

secret
The secret of the Amazon S3 bucket. Defaults to the AWS_SECRET_ACCESS_KEY environment variable.

cache
The local folder to use as a cache, defaults to board_cache_path().

host
The host to use for storage, defaults to "s3.amazonaws.com".

...
Additional parameters required to initialize a particular board.

Details

This function requires an Amazon S3 bucket to be manually created; otherwise, registering an S3 board will fail.

See also

Examples

if (FALSE) {
# the following example requires an Amazon S3 API key
board_register_s3(bucket = "s3bucket")
}