Pin Resource

Pins the given resource locally or to the given board.

pin(x, name = NULL, description = NULL, board = NULL, ...)

Arguments

x
An object, local file or remote URL to pin.

name
The name for the dataset or object.

description
Optional description for this pin.

board
The board where this pin will be placed.

...
Additional parameters.

Details

pin() allows you to cache remote resources and intermediate results with ease. When caching remote resources, usually URLs, it will check for HTTP caching headers to avoid re-downloading when the remote result has not changed.

This makes it ideal to support reproducible research by requiring manual instruction to download resources before running your R script.

In addition, pin() still works when working offline or when the remote resource becomes unavailable; when this happens, a warning will be triggered but your code will continue to work.

Examples

library(pins)

# define local board
board_register_local(cache = tempfile())

# cache the mtcars dataset
pin(mtcars)

# cache computation over mtcars
mtcars[mtcars$mpg > 30,] %>%
  pin(name = "mtefficient")

# retrieve cached pin
pin_get("mtefficient")

#> # A tibble: 4 x 11
#>     mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1  32.4     4  78.7    66  4.08  2.2   19.5     1     1     4     1
#> 2  30.4     4  75.7    52  4.93  1.62  18.5     1     1     4     2
#> 3  33.9     4  71.1    65  4.22  1.84  19.9     1     1     4     1
#> 4  30.4     4  95.1   113  3.77  1.51  16.9     1     1     5     2

# url to remote resource
resource <- file.path("https://raw.githubusercontent.com/facebook/prophet",
                      "master/examples/example_retail_sales.csv")

# cache remote resource
pin(resource, name = "example_retail_sales")

#> Error in if (file.exists(data_txt)) {    manifest <- yaml::read_yaml(data_txt, eval.expr = FALSE)}: argument is of length zero

# load cached csv
pin_get("example_retail_sales") %>% read.csv()

#> Error in pin_get("example_retail_sales"): Failed to retrieve 'example_retail_sales' pin.

# cache and read csv
read.csv(pin(resource))

#> Error in if (file.exists(data_txt)) {    manifest <- yaml::read_yaml(data_txt, eval.expr = FALSE)}: argument is of length zero