Browse source of a pin

pin_browse() navigates you to the home of a pin, either on the internet or on your local file system.

Usage

pin_browse(board, name, version = NULL, local = FALSE)

Arguments

board

A pin board, created by board_folder(), board_connect(), board_url() or another board_ function.

name

Pin name.

version

Retrieve a specific version of a pin. Use pin_versions() to find out which versions are available and when they were created.

local

If TRUE, will open the local copy of the pin; otherwise will show you the home of the pin on the internet.

Examples

board <- board_temp(versioned = TRUE)
board |> pin_write(1:10, "x")
#> Guessing `type = 'rds'`
#> Creating new version '20250904T024501Z-23075'
#> Writing to pin 'x'
board |> pin_write(1:11, "x")
#> Guessing `type = 'rds'`
#> Creating new version '20250904T024501Z-215e0'
#> Writing to pin 'x'
board |> pin_write(1:12, "x")
#> Guessing `type = 'rds'`
#> Creating new version '20250904T024501Z-91f1b'
#> Writing to pin 'x'

board |> pin_browse("x", local = TRUE)
#> ℹ Pin at </tmp/Rtmp5t4hSz/pins-1e0c4ac69872/x/20250904T024501Z-91f1b>