Upload and download files to and from a board

This is a lower-level interface than pin_read() and pin_write() that you can use to pin any file.

pin_download(board, name, version = NULL, hash = NULL, ...)

pin_upload(board, paths, name = NULL, desc = NULL, metadata = NULL, ...)

Arguments

board
A pin board, created by board_folder(), board_rsconnect(), board_url() or other other board_ function.

name
Pin name.

version
For boards that support versions, this allows you to retrieve a specific version of a pin. See vignette("versioning)" for details.

hash
Specify a hash to verify that you get exactly the dataset that you expect. You can find the hash of an existing pin by looking for pin_hash in pin_meta().

...
Additional arguments passed on to methods for a specific board.

paths
A character vector of file paths to upload to board.

desc
A text description of the pin; most important for shared boards so that others can understand what the pin contains. If omitted, pins will generate a brief description of the contents.

metadata
A list containing additional metadata to store with the pin. When retrieving the pin, this will be stored in the user key, to avoid potential clashes with the metadata that pins itself uses.

Value

pin_download() returns a character vector of file paths; pin_upload() returns board, invisibly.

Examples

board <- board_temp()

board %>% pin_upload(system.file("CITATION"))
#> Guessing `name = 'CITATION'`
#> Creating new version '20210521T125658Z-dec01'
path <- board %>% pin_download("CITATION")
path
#> [1] "/tmp/RtmpdOC0ZQ/pins-296b7a20f72f/CITATION/20210521T125658Z-dec01/CITATION"
readLines(path)[1:5]
#> [1] "bibentry(\"Manual\","                                                         
#> [2] "         title = \"R: A Language and Environment for Statistical Computing\","
#> [3] "         author = person(\"R Core Team\"),"                                   
#> [4] "         organization = \"R Foundation for Statistical Computing\","          
#> [5] "         address      = \"Vienna, Austria\","