Find Pin

Find a pin in any board registered using board_register().

pin_find(text = NULL, board = NULL, name = NULL, extended = FALSE,
  ...)

Arguments

text
The text to find in the pin description or name.

board
The board name used to find the pin.

name
The exact name of the pin to match when searching.

extended
Should additional board-specific columns be shown?

...
Additional parameters.

Details

pin_find() allows you to discover new resources or retrieve pins you've previously created with pin().

The pins package comes with a CRAN packages board which allows searching all CRAN packages; however, you can add additional boards to search from like Kaggle, Github and RStudio Connect.

For 'local' and 'packages' boards, the 'text' parameter searches the title and description of a pin using a regular expression. Other boards search in different ways, most of them are just partial matches, please refer to their documentation to understand how other boards search for pins.

Once you find a pin, you can retrieve with pin_get("pin-name").

Examples

library(pins)

# retrieve pins
pin_find()

#> # A tibble: 18,940 x 4
#>    name           description                                      type  board  
#>    <chr>          <chr>                                            <chr> <chr>  
#>  1 mtcars         ""                                               table temp   
#>  2 mtefficient    ""                                               table temp   
#>  3 iris           "This famous (Fisher's or Anderson's) iris data… table txtexa…
#>  4 mtcars         "The data was extracted from the 1974 Motor Tre… table txtexa…
#>  5 A3/housing     "Boston Housing Prices from A3 package."         table packag…
#>  6 A3/multifunct… "Ecosystem Multifunctionality from A3 package."  table packag…
#>  7 aaSEA/AAindex  "A data frame of 533 amino acid properties from… table packag…
#>  8 aaSEA/corSubF… "A data frame of correlated sites from aaSEA pa… table packag…
#>  9 aaSEA/Cruciani "A data frame of 3 Cruciani properties for 20 a… table packag…
#> 10 aaSEA/Fasgai   "A data frame of six Fasgai vectors for 20 amin… table packag…
#> # … with 18,930 more rows

# search pins related to 'cars'
pin_find("cars")

#> # A tibble: 40 x 4
#>    name             description                                    type  board  
#>    <chr>            <chr>                                          <chr> <chr>  
#>  1 mtcars           ""                                             table temp   
#>  2 mtcars           "The data was extracted from the 1974 Motor T… table txtexa…
#>  3 bestNormalize/a… "Prices of 6,283 cars listed on Autotrader fr… table packag…
#>  4 BSDA/Jdpower     "Number of problems reported per 100 cars in … table packag…
#>  5 BSDA/Marked      "Percent of marked cars in 65 police departme… table packag…
#>  6 BSDA/Supercar    "Top speeds attained by five makes of superca… table packag…
#>  7 caret/cars       "Kelly Blue Book resale data for 2005 model y… table packag…
#>  8 CornerstoneR/ca… "Data from carstats from CornerstoneR package… table packag…
#>  9 DAAG/Cars93.sum… "A Summary of the Cars93 Data set from DAAG p… table packag…
#> 10 DAAG/modelcars   "Model Car Data from DAAG package."            table packag…
#> # … with 30 more rows

# search pins related to 'seattle' in the 'packages' board
pin_find("seattle", board = "packages")

#> # A tibble: 6 x 4
#>   name               description                                    type  board 
#>   <chr>              <chr>                                          <chr> <chr> 
#> 1 hpiR/ex_sales      Subset of Seattle Home Sales from hpiR packag… table packa…
#> 2 hpiR/seattle_sales Seattle Home Sales from hpiR package.          table packa…
#> 3 latticeExtra/Seat… Daily Rainfall and Temperature at the Seattle… table packa…
#> 4 microsynth/seattl… Data for a crime intervention in Seattle, Was… table packa…
#> 5 vegawidget/data_s… Example dataset: Seattle daily weather from v… table packa…
#> 6 vegawidget/data_s… Example dataset: Seattle hourly temperatures … table packa…

# search pins related to 'london' in the 'packages' board
pin_find("london", board = "packages")

#> # A tibble: 30 x 4
#>    name                     description                             type  board 
#>    <chr>                    <chr>                                   <chr> <chr> 
#>  1 bsamGP/London.Mortality  Daily Moratlity in London from bsamGP … table packa…
#>  2 DAAG/poxetc              Deaths from various causes, in London … table packa…
#>  3 DAAG/whoops              Deaths from whooping cough, in London … table packa…
#>  4 epimdr/meas              Bi-weekly measles incidence in London … table packa…
#>  5 ev.trawl/hourly_bloomsb… Hourly measurements of 6 air pollutant… table packa…
#>  6 extremis/lse             Selected Stocks from the London Stock … table packa…
#>  7 geofacet/london_afford   london_afford from geofacet package.    table packa…
#>  8 HistData/Arbuthnot       Arbuthnot's data on male and female bi… table packa…
#>  9 HistData/Cholera         William Farr's Data on Cholera in Lond… table packa…
#> 10 jomo/ExamScores          Exam results for six inner London Educ… table packa…
#> # … with 20 more rows

# \donttest{
# retrieve 'hpiR/seattle_sales' pin
pin_get("hpiR/seattle_sales")

#> # A tibble: 43,313 x 16
#>    pinx  sale_id sale_price sale_date  use_type  area lot_sf  wfnt bldg_grade
#>    <chr> <chr>        <int> <date>     <chr>    <int>  <int> <dbl>      <int>
#>  1 ..00… 2013..…     289000 2013-02-06 sfr         79   9295     0          7
#>  2 ..00… 2013..…     356000 2013-07-11 sfr         18   6000     0          6
#>  3 ..00… 2010..…     333500 2010-12-29 sfr         79   7200     0          8
#>  4 ..00… 2016..…     577200 2016-03-17 sfr         79   7200     0          8
#>  5 ..00… 2012..…     237000 2012-05-02 sfr         79   5662     0          7
#>  6 ..00… 2014..…     347500 2014-03-11 sfr         79   5830     0          7
#>  7 ..00… 2012..…     429000 2012-09-20 sfr         18  12700     0          7
#>  8 ..00… 2015..…     653295 2015-07-21 sfr         79   7000     0          7
#>  9 ..00… 2014..…     427650 2014-02-19 townhou…    79   3072     0          7
#> 10 ..00… 2015..…     488737 2015-03-19 townhou…    79   3072     0          7
#> # … with 43,303 more rows, and 7 more variables: tot_sf <int>, beds <int>,
#> #   baths <dbl>, age <int>, eff_age <int>, longitude <dbl>, latitude <dbl>

# retrieve 'bsamGP/London.Mortality' pin
pin_get("bsamGP/London.Mortality")

#> # A tibble: 5,113 x 7
#>    date        tmean  tmin  tmax   dewp    rh death
#>    <fct>       <dbl> <dbl> <dbl>  <dbl> <dbl> <int>
#>  1 1993-01-01  4.63   1.57  7.67  2.39   79.4   233
#>  2 1993-01-02 -0.168 -2.76  2.40 -1.50   85.9   228
#>  3 1993-01-03 -0.875 -5.16  3.40 -5.22   69.4   228
#>  4 1993-01-04  2.48  -1.99  6.94  2.04   88.2   223
#>  5 1993-01-05  7.45   3.45 11.4   8.63   94.0   246
#>  6 1993-01-06  9.95   7.78 12.1  10.1    90.0   242
#>  7 1993-01-07  6.98   2.92 11.0   5.93   93.6   226
#>  8 1993-01-08  7.34   3.43 11.2   0.177  67.0   218
#>  9 1993-01-09  7.91   3.65 12.2   8.93   84.8   212
#> 10 1993-01-10  9.92   7.79 12.1   9.63   89.4   218
#> # … with 5,103 more rows

# }