Retrieve pin metadata (legacy API)

Retrieve metadata for pins in legacy boards.

pin_info(
  name,
  board = NULL,
  extended = TRUE,
  metadata = TRUE,
  signature = FALSE,
  ...
)

Arguments

name
The exact name of the pin to match when searching.

board
The board name used to find the pin.

extended
Should additional board-specific information be shown?

metadata
Should additional pin-specific information be shown?

signature
Should a signature to identify this pin be shown?

...
Additional parameters.

Examples

# old API
board_register_local(cache = tempfile())
pin(mtcars)
pin_info("mtcars", "local")
#> Warning: `pin_info()` was deprecated in pins 1.0.0.
#> Please use `pin_meta()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.
#> # Source: local<mtcars> [table]
#> # Properties:
#> #   rows: 32
#> #   cols: 11
#> #   columns:
#> #     mpg: numeric
#> #     cyl: numeric
#> #     disp: numeric
#> #     hp: numeric
#> #     drat: numeric
#> #     wt: numeric
#> #     qsec: numeric
#> #     vs: numeric
#> #     am: numeric
#> #     gear: numeric
#> #     carb: numeric

# new API
board <- board_temp()
board %>% pin_write(mtcars)
#> Using `name = 'mtcars'`
#> Guessing `type = 'rds'`
#> Creating new version '20210824T215735Z-5a338'
board %>% pin_meta("mtcars")
#> List of 9
#>  $ file       : chr "mtcars.rds"
#>  $ file_size  : 'fs_bytes' int 1.19K
#>  $ pin_hash   : chr "5a338159141e2a0b"
#>  $ type       : chr "rds"
#>  $ description: chr "A pin containing a data frame with 32 rows and 11 columns"
#>  $ created    : POSIXct[1:1], format: "2021-08-24 21:57:00"
#>  $ api_version: num 1
#>  $ user       : list()
#>  $ local      :List of 2
#>   ..$ dir    : 'fs_path' chr "/tmp/Rtmp41A4UU/pins-287f40792b45/mtcars/20210824T215735Z-5a338"
#>   ..$ version: chr "20210824T215735Z-5a338"