Retrieve metadata for a pin

Retrieve metadata for a pin

pin_meta(board, name, version = NULL, ...)

Arguments

board
A pin board, created by board_folder(), board_rsconnect(), board_github() or other other board_ function.

name
Pin name.

version
For boards that support versions, this allows you to retrieve a specific version of a pin. See vignette("versioning)" for details.

...
Additional arguments passed on to methods for a specific board.

Value

A list.

Examples

b <- board_temp()
b %>% pin_write(head(mtcars), "mtcars", metadata = list("Hadley" = TRUE))

#> Guessing `type = 'rds'`

#> Creating new version 'f879756e28a960ae'

# Get the pin
b %>% pin_read("mtcars")

#>                    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4         21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag     21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710        22.8   4  108  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive    21.4   6  258 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout 18.7   8  360 175 3.15 3.440 17.02  0  0    3    2
#> Valiant           18.1   6  225 105 2.76 3.460 20.22  1  0    3    1

# Get its metadata
b %>% pin_meta("mtcars")

#> List of 10
#>  $ file       : chr "mtcars.rds"
#>  $ file_size  : int 426
#>  $ pin_hash   : chr "f879756e28a960ae"
#>  $ type       : chr "rds"
#>  $ description: chr "A pin containing a data frame with 6 rows and 11 columns"
#>  $ created    : chr "2021-04-26T22:17:58"
#>  $ api_version: num 1
#>  $ user       :List of 1
#>   ..$ Hadley: logi TRUE
#>  $ version    : chr "f879756e28a960ae"
#>  $ pin_path   : 'fs_path' chr "/tmp/RtmpyvUPpp/pins-2dc1dc88935/mtcars/f879756e28a960ae"

# Get path to underlying data
b %>% pin_download("mtcars")

#> [1] "/tmp/RtmpyvUPpp/pins-2dc1dc88935/mtcars/f879756e28a960ae/mtcars.rds"