Search for pins

The underlying search method depends on the board, but most will search for text in the pin name and description.

pin_search(board, pattern = NULL)

Arguments

board
A pin board, created by board_folder(), board_rsconnect(), board_github() or other other board_ function.

pattern
A string to search for in pin name and description.

Value

A data frame that summarises the metadata for each pin. Key columns (version, type, description, created, and file_size) are pulled out into columns; everything else can be found in the meta list-column.

Examples

board <- board_temp()

board %>% pin_write(1:5, "x", desc = "Some numbers")

#> Guessing `type = 'rds'`

#> Creating new version '4cbcc49c50fda258'

board %>% pin_write(letters[c(1, 5, 10, 15, 21)], "y", desc = "My favourite letters")

#> Guessing `type = 'rds'`

#> Creating new version 'f807f38ce47982d5'

board %>% pin_write(runif(20), "z", desc = "Random numbers")

#> Guessing `type = 'rds'`

#> Creating new version '15c6d38de0ebb0fb'

board %>% pin_search()

#> # A tibble: 3 x 7
#>   name  version type  description        created              file_size meta    
#>   <chr>   <int> <chr> <chr>              <dttm>              <fs::byte> <list>  
#> 1 x           1 rds   Some numbers       2021-04-22 22:11:10         50 <pins_m…
#> 2 y           1 rds   My favourite lett… 2021-04-22 22:11:10         56 <pins_m…
#> 3 z           1 rds   Random numbers     2021-04-22 22:11:10        169 <pins_m…

board %>% pin_search("number")

#> # A tibble: 2 x 7
#>   name  version type  description    created               file_size meta      
#>   <chr>   <int> <chr> <chr>          <dttm>              <fs::bytes> <list>    
#> 1 x           1 rds   Some numbers   2021-04-22 22:11:10          50 <pins_met>
#> 2 z           1 rds   Random numbers 2021-04-22 22:11:10         169 <pins_met>

board %>% pin_search("letters")

#> # A tibble: 1 x 7
#>   name  version type  description        created              file_size meta    
#>   <chr>   <int> <chr> <chr>              <dttm>              <fs::byte> <list>  
#> 1 y           1 rds   My favourite lett… 2021-04-22 22:11:10         56 <pins_m…