Use RStudio Connect as board

To use a RStudio Connect board, you need to first authenticate. The easiest way to do so is by launching Tools - Global Options - Publishing - Connect, and follow the instructions.

You can share pins with others in RStudio Connect by changing the viewers of the document to specific users or groups. This is accomplished by opening the new published pin and then changing access under the settings tab. After you've shared the pin, it will be automatically available to others.

Usage

board_rsconnect(
  auth = c("auto", "manual", "envvar", "rsconnect"),
  server = NULL,
  account = NULL,
  key = NULL,
  output_files = FALSE,
  cache = NULL,
  name = "rsconnect",
  versioned = TRUE,
  use_cache_on_failure = is_interactive(),
  versions = deprecated()
)

Arguments

auth

There are three ways to authenticate:

auth = "manual" uses arguments server and key.

auth = "envvar" uses environment variables CONNECT_API_KEY and CONNECT_SERVER.

auth = "rsconnect" uses servers registered with the rsconnect package (filtered by server and account, if provided)

The default, auth = "auto", automatically picks between the three options, using "manual" if server and key are provided, "envvar" if both environment variables are set, and "rsconnect" otherwise.

server

For auth = "manual" or auth = 'envvar', the full url to the server, like http://server.rstudio.com/rsc or https://connect.rstudio.com/. For auth = 'rsconnect' a host name used to disambiguate RSC accounts, like server.rstudio.com or connect.rstudio.com.

account

A user name used to disambiguate multiple RSC accounts.

key

The RStudio Connect API key.

output_files

No longer supported.

cache

Cache path. Every board requires a local cache to avoid downloading files multiple times. The default stores in a standard cache location for your operating system, but you can override if needed.

name

An optional name used identify the board. This is no longer generally needed since you should be passing around an explicit board object.

versioned

Should this board be registered with support for versions?

use_cache_on_failure

If the pin fails to download, is it ok to use the last cached version? Defaults to is_interactive() so you'll be robust to poor internet connectivity when exploring interactively, but you'll get clear errors when the code is deployed.

versions

Should this board be registered with support for versions?

Public pins

If your RSC instance allows it, you can share a pin publicly by setting the access type to all:

board %>% pin_write(my_df, access_type = "all")

(You can also do this in RSC by setting "Access" to "Anyone - no login required")

Now anyone can read your pin through board_url():

board <- board_url(c(
  my_df = "https://connect.rstudioservices.com/content/3004/"
))
board %>% pin_read("my_df")

You can find the URL of a pin with pin_browse().

See also

Examples

if (FALSE) {
board <- board_rsconnect()
# Share the mtcars with your team
board %>% pin_write(mtcars, "mtcars")

# Download a shared dataset
board %>% pin_read("timothy/mtcars")
}