Use an RStudio Connect board

To use a RStudio Connect board, you need to first authenticate. The easiest way to do so is by launching Tools - Global Options - Publishing - Connect, and follow the instructions.

You can share pins with others in RStudio Connect by changing the viewers of the document to specific users or groups. This is accomplished by opening the new published pin and then changing access under the settings tab. After you've shared the pin, it will be automatically available to others.

board_rsconnect(
  auth = c("auto", "envvar", "rsconnect"),
  server = NULL,
  account = NULL,
  key = NULL,
  output_files = FALSE,
  cache = NULL,
  name = "rsconnect",
  versions = TRUE,
  ...
)

Arguments

auth
There are two approaches to auth: you can either use "envvars" CONNECT_API_KEY and CONNECT_SERVER or the rsconnect package. The default is auto, which will use the environment variables if both are available, and rsconnect if not.

server
For auth = "envvar" the full url to the server. For auth = 'rsconnect' a host name used to disambiguate RSC accounts.

account
A user name used to disambiguate multiple RSC accounts

key
The RStudio Connect API key.

output_files
`r lifecycle::badge("deprecated") No longer supported.

cache
Cache path. Every board requires a local cache to avoid downloading files multiple times. The default stores in a standard cache location for your operating system, but you can override if needed.

name
An optional name used identify the board. This is no longer generally needed since you should be passing around an explicit board object.

versions
Should this board be registered with support for versions?

...
Additional parameters required to initialize a particular board.

Public pins

If your RSC instance allows it, you can share a pin publicly by setting the access type to all:

board %>% pin_write(my_df, access_type = "all")

(You can also do this in RSC by setting "Access" to "Anyone - no login required")

Now anyone can read your pin through board_url():

board <- board_url(c(
  my_df = "https://connect.rstudioservices.com/content/3004/"
))
board %>% pin_read("my_df")

You can find the URL of a pin with pin_browse().

See also

Examples

if (FALSE) {
board <- board_rsconnect()
# Share the mtcars with your team
board %>% pin_write(mtcars, "mtcars")

# Download a shared dataset
board %>% pin_read("timothy/mtcars")
}