Find Pin

Find a pin in any board registered using board_register().

pin_find(text = NULL, board = NULL, name = NULL, extended = FALSE, ...)

Arguments

text
The text to find in the pin description or name.

board
The board name used to find the pin.

name
The exact name of the pin to match when searching.

extended
Should additional board-specific columns be shown?

...
Additional parameters.

Details

pin_find() allows you to discover new resources or retrieve pins you've previously created with pin().

The pins package comes with a CRAN packages board which allows searching all CRAN packages; however, you can add additional boards to search from like Kaggle, Github and RStudio Connect.

For 'local' and 'packages' boards, the 'text' parameter searches the title and description of a pin using a regular expression. Other boards search in different ways, most of them are just partial matches, please refer to their documentation to understand how other boards search for pins.

Once you find a pin, you can retrieve with pin_get("pin-name").

Examples

library(pins)

# retrieve pins
pin_find()

#> # A tibble: 5 x 4
#>   name           description                                       type  board  
#>   <chr>          <chr>                                             <chr> <chr>  
#> 1 example_retai… ""                                                files local  
#> 2 iris           "This famous (Fisher's or Anderson's) iris data … table txtexa…
#> 3 mtcars         ""                                                table local  
#> 4 mtcars         "The data was extracted from the 1974 Motor Tren… table txtexa…
#> 5 mtefficient    ""                                                table local  

# search pins related to 'cars'
pin_find("cars")

#> # A tibble: 2 x 4
#>   name   description                                              type  board   
#>   <chr>  <chr>                                                    <chr> <chr>   
#> 1 mtcars ""                                                       table local   
#> 2 mtcars "The data was extracted from the 1974 Motor Trend US ma… table txtexam…

# search pins related to 'seattle' in the 'packages' board
pin_find("seattle", board = "packages")

#> # A tibble: 6 x 4
#>   name               description                                    type  board 
#>   <chr>              <chr>                                          <chr> <chr> 
#> 1 hpiR/ex_sales      Subset of Seattle Home Sales from hpiR packag… table packa…
#> 2 hpiR/seattle_sales Seattle Home Sales from hpiR package.          table packa…
#> 3 latticeExtra/Seat… Daily Rainfall and Temperature at the Seattle… table packa…
#> 4 microsynth/seattl… Data for a crime intervention in Seattle, Was… table packa…
#> 5 vegawidget/data_s… Example dataset: Seattle daily weather from v… table packa…
#> 6 vegawidget/data_s… Example dataset: Seattle hourly temperatures … table packa…

# search pins related to 'london' in the 'packages' board
pin_find("london", board = "packages")

#> # A tibble: 30 x 4
#>    name                     description                             type  board 
#>    <chr>                    <chr>                                   <chr> <chr> 
#>  1 bsamGP/London.Mortality  Daily Moratlity in London from bsamGP … table packa…
#>  2 DAAG/poxetc              Deaths from various causes, in London … table packa…
#>  3 DAAG/whoops              Deaths from whooping cough, in London … table packa…
#>  4 epimdr/meas              Bi-weekly measles incidence in London … table packa…
#>  5 ev.trawl/hourly_bloomsb… Hourly measurements of 6 air pollutant… table packa…
#>  6 extremis/lse             Selected Stocks from the London Stock … table packa…
#>  7 geofacet/london_afford   london_afford from geofacet package.    table packa…
#>  8 HistData/Arbuthnot       Arbuthnot's data on male and female bi… table packa…
#>  9 HistData/Cholera         William Farr's Data on Cholera in Lond… table packa…
#> 10 jomo/ExamScores          Exam results for six inner London Educ… table packa…
#> # … with 20 more rows