Find Pin

Find a pin in any board registered using board_register().

pin_find(text = NULL, board = NULL, name = NULL, ...)

Arguments

text
The text to find in the pin description or name.

board
The board name used to find the pin.

name
The exact name of the pin to match when searching.

...
Additional parameters.

Details

pin_find() allows you to discover new resources or retrieve pins you've previously created with pin().

The pins package comes with a CRAN packages board which allows searching all CRAN packages; however, you can add additional boards to search from like Kaggle, Github and RStudio Connect.

For 'local' and 'packages' boards, the 'text' parameter searches the title and description of a pin using a regular expression. Other boards search in different ways, most of them are just partial matches, please refer to their documentation to understand how other boards search for pins.

Once you find a pin, you can retrieve with pin_get("pin-name").

Examples

library(pins)

# retrieve pins
pin_find()

#> # A tibble: 18,700 x 4
#>    name            description                                     type  board  
#>    <chr>           <chr>                                           <chr> <chr>  
#>  1 iris            This famous (Fisher's or Anderson's) iris data… table txtexa…
#>  2 mtcars          The data was extracted from the 1974 Motor Tre… table txtexa…
#>  3 mtcars          ""                                              table local  
#>  4 mtefficient     ""                                              table local  
#>  5 example_retail… ""                                              files local  
#>  6 A3/housing      Boston Housing Prices from A3 package.          table packag…
#>  7 A3/multifuncti… Boston Housing Prices from A3 package.          table packag…
#>  8 abc.data/musig… Boston Housing Prices from abc.data package.    table packag…
#>  9 abc.data/ppc    Boston Housing Prices from abc.data package.    table packag…
#> 10 ABC.RAP/annota… Boston Housing Prices from ABC.RAP package.     table packag…
#> # … with 18,690 more rows

# search pins related to 'cars'
pin_find("cars")

#> # A tibble: 30 x 4
#>    name           description                                      type  board  
#>    <chr>          <chr>                                            <chr> <chr>  
#>  1 mtcars         The data was extracted from the 1974 Motor Tren… table txtexa…
#>  2 mtcars         ""                                               table local  
#>  3 CornerstoneR/… Boston Housing Prices from CornerstoneR package. table packag…
#>  4 DAAG/Cars93.s… Boston Housing Prices from DAAG package.         table packag…
#>  5 DAAG/modelcars Boston Housing Prices from DAAG package.         table packag…
#>  6 DAAG/toycars   Boston Housing Prices from DAAG package.         table packag…
#>  7 dobson/Cars    Boston Housing Prices from dobson package.       table packag…
#>  8 easyalluvial/… Boston Housing Prices from easyalluvial package. table packag…
#>  9 ISLR/Carseats  Boston Housing Prices from ISLR package.         table packag…
#> 10 lessR/dataCar… Boston Housing Prices from lessR package.        table packag…
#> # … with 20 more rows

# search pins related to 'seattle' in the 'packages' board
pin_find("seattle", board = "packages")

#> # A tibble: 4 x 4
#>   name                      description                            type  board  
#>   <chr>                     <chr>                                  <chr> <chr>  
#> 1 hpiR/seattle_sales        Boston Housing Prices from hpiR packa… table packag…
#> 2 microsynth/seattledmi     Boston Housing Prices from microsynth… table packag…
#> 3 vegawidget/data_seattle_… Boston Housing Prices from vegawidget… table packag…
#> 4 vegawidget/data_seattle_… Boston Housing Prices from vegawidget… table packag…

# search pins related to 'london' in the 'packages' board
pin_find("london", board = "packages")

#> # A tibble: 6 x 4
#>   name                             description                      type  board 
#>   <chr>                            <chr>                            <chr> <chr> 
#> 1 bsamGP/London.Mortality          Boston Housing Prices from bsam… table packa…
#> 2 geofacet/london_afford           Boston Housing Prices from geof… table packa…
#> 3 MultiplierDEA/Metropolitan_And_… Boston Housing Prices from Mult… table packa…
#> 4 OIdata/london_boroughs           Boston Housing Prices from OIda… table packa…
#> 5 rmweather/data_london_normalised Boston Housing Prices from rmwe… table packa…
#> 6 rmweather/data_london            Boston Housing Prices from rmwe… table packa…

# \donttest{
# retrieve 'hpiR/seattle_sales' pin
pin_get("hpiR/seattle_sales")

#> # A tibble: 43,313 x 16
#>    pinx  sale_id sale_price sale_date  use_type  area lot_sf  wfnt bldg_grade
#>    <chr> <chr>        <int> <date>     <chr>    <int>  <int> <dbl>      <int>
#>  1 ..00… 2013..…     289000 2013-02-06 sfr         79   9295     0          7
#>  2 ..00… 2013..…     356000 2013-07-11 sfr         18   6000     0          6
#>  3 ..00… 2010..…     333500 2010-12-29 sfr         79   7200     0          8
#>  4 ..00… 2016..…     577200 2016-03-17 sfr         79   7200     0          8
#>  5 ..00… 2012..…     237000 2012-05-02 sfr         79   5662     0          7
#>  6 ..00… 2014..…     347500 2014-03-11 sfr         79   5830     0          7
#>  7 ..00… 2012..…     429000 2012-09-20 sfr         18  12700     0          7
#>  8 ..00… 2015..…     653295 2015-07-21 sfr         79   7000     0          7
#>  9 ..00… 2014..…     427650 2014-02-19 townhou…    79   3072     0          7
#> 10 ..00… 2015..…     488737 2015-03-19 townhou…    79   3072     0          7
#> # … with 43,303 more rows, and 7 more variables: tot_sf <int>, beds <int>,
#> #   baths <dbl>, age <int>, eff_age <int>, longitude <dbl>, latitude <dbl>

# retrieve 'bsamGP/London.Mortality' pin
pin_get("bsamGP/London.Mortality")

#> # A tibble: 5,113 x 7
#>    date        tmean  tmin  tmax   dewp    rh death
#>    <fct>       <dbl> <dbl> <dbl>  <dbl> <dbl> <int>
#>  1 1993-01-01  4.63   1.57  7.67  2.39   79.4   233
#>  2 1993-01-02 -0.168 -2.76  2.40 -1.50   85.9   228
#>  3 1993-01-03 -0.875 -5.16  3.40 -5.22   69.4   228
#>  4 1993-01-04  2.48  -1.99  6.94  2.04   88.2   223
#>  5 1993-01-05  7.45   3.45 11.4   8.63   94.0   246
#>  6 1993-01-06  9.95   7.78 12.1  10.1    90.0   242
#>  7 1993-01-07  6.98   2.92 11.0   5.93   93.6   226
#>  8 1993-01-08  7.34   3.43 11.2   0.177  67.0   218
#>  9 1993-01-09  7.91   3.65 12.2   8.93   84.8   212
#> 10 1993-01-10  9.92   7.79 12.1   9.63   89.4   218
#> # … with 5,103 more rows

# }