Retrieve pin metadata (legacy API)

Retrieve metadata for pins in legacy boards.

Usage

pin_info(
  name,
  board = NULL,
  extended = TRUE,
  metadata = TRUE,
  signature = FALSE,
  ...
)

Arguments

name

The exact name of the pin to match when searching.

board

The board name used to find the pin.

extended

Should additional board-specific information be shown?

metadata

Should additional pin-specific information be shown?

signature

Should a signature to identify this pin be shown?

...

Additional parameters.

Examples

# old API
board_register_local(cache = tempfile())
pin(mtcars)
pin_info("mtcars", "local")
#> # Source: local<mtcars> [table]
#> # Properties:
#> #   rows: 32
#> #   cols: 11
#> #   columns:
#> #     mpg: numeric
#> #     cyl: numeric
#> #     disp: numeric
#> #     hp: numeric
#> #     drat: numeric
#> #     wt: numeric
#> #     qsec: numeric
#> #     vs: numeric
#> #     am: numeric
#> #     gear: numeric
#> #     carb: numeric

# new API
board <- board_temp()
board %>% pin_write(mtcars)
#> Using `name = 'mtcars'`
#> Guessing `type = 'rds'`
#> Creating new version '20240802T172558Z-b9134'
#> Writing to pin 'mtcars'
board %>% pin_meta("mtcars")
#> List of 13
#>  $ file       : chr "mtcars.rds"
#>  $ file_size  : 'fs_bytes' int 1.19K
#>  $ pin_hash   : chr "b91346fd7a21419e"
#>  $ type       : chr "rds"
#>  $ title      : chr "mtcars: a pinned 32 x 11 data frame"
#>  $ description: NULL
#>  $ tags       : NULL
#>  $ urls       : NULL
#>  $ created    : POSIXct[1:1], format: "2024-08-02 17:25:58"
#>  $ api_version: int 1
#>  $ user       : list()
#>  $ name       : chr "mtcars"
#>  $ local      :List of 3
#>   ..$ dir    : 'fs_path' chr "/tmp/RtmpECtVjC/pins-16cc7f8f84fd/mtcars/20240802T172558Z-b9134"
#>   ..$ url    : NULL
#>   ..$ version: chr "20240802T172558Z-b9134"