Retrieve pin metadata (legacy API)

Retrieve metadata for pins in legacy boards.

Usage

pin_info(
  name,
  board = NULL,
  extended = TRUE,
  metadata = TRUE,
  signature = FALSE,
  ...
)

Arguments

name

The exact name of the pin to match when searching.

board

The board name used to find the pin.

extended

Should additional board-specific information be shown?

metadata

Should additional pin-specific information be shown?

signature

Should a signature to identify this pin be shown?

...

Additional parameters.

Examples

# old API
board_register_local(cache = tempfile())
pin(mtcars)
pin_info("mtcars", "local")
#> # Source: local<mtcars> [table]
#> # Properties:
#> #   rows: 32
#> #   cols: 11
#> #   columns:
#> #     mpg: numeric
#> #     cyl: numeric
#> #     disp: numeric
#> #     hp: numeric
#> #     drat: numeric
#> #     wt: numeric
#> #     qsec: numeric
#> #     vs: numeric
#> #     am: numeric
#> #     gear: numeric
#> #     carb: numeric

# new API
board <- board_temp()
board %>% pin_write(mtcars)
#> Using `name = 'mtcars'`
#> Guessing `type = 'rds'`
#> Creating new version '20230314T164553Z-76dea'
#> Writing to pin 'mtcars'
board %>% pin_meta("mtcars")
#> List of 12
#>  $ file       : chr "mtcars.rds"
#>  $ file_size  : 'fs_bytes' int 1.19K
#>  $ pin_hash   : chr "76dea67ad2bfdae7"
#>  $ type       : chr "rds"
#>  $ title      : chr "mtcars: a pinned 32 x 11 data frame"
#>  $ description: NULL
#>  $ tags       : NULL
#>  $ created    : POSIXct[1:1], format: "2023-03-14 16:45:53"
#>  $ api_version: num 1
#>  $ user       : list()
#>  $ name       : chr "mtcars"
#>  $ local      :List of 3
#>   ..$ dir    : 'fs_path' chr "/tmp/Rtmpv9beOX/pins-18226701f6d1/mtcars/20230314T164553Z-76dea"
#>   ..$ url    : NULL
#>   ..$ version: chr "20230314T164553Z-76dea"