Pin Info

pin_info() is deprecated in favour of pin_meta().

pin_info(
  name,
  board = NULL,
  extended = TRUE,
  metadata = TRUE,
  signature = FALSE,
  ...
)

Arguments

name
The exact name of the pin to match when searching.

board
The board name used to find the pin.

extended
Should additional board-specific information be shown?

metadata
Should additional pin-specific information be shown?

signature
Should a signature to identify this pin be shown?

...
Additional parameters.

Examples

board <- board_temp()
pin(mtcars, board = board)

pin_info("mtcars", board = board)

#> Warning: `pin_info()` was deprecated in pins 1.0.0.
#> Please use `pin_meta()` instead.

#> # Source: temp<mtcars> [table]
#> # Properties:
#> #   rows: 32
#> #   cols: 11
#> #   columns:
#> #     mpg: numeric
#> #     cyl: numeric
#> #     disp: numeric
#> #     hp: numeric
#> #     drat: numeric
#> #     wt: numeric
#> #     qsec: numeric
#> #     vs: numeric
#> #     am: numeric
#> #     gear: numeric
#> #     carb: numeric

# ->
board %>% pin_meta("mtcars")

#> List of 8
#>  $ type       : chr "table"
#>  $ description: NULL
#>  $ rows       : int 32
#>  $ cols       : int 11
#>  $ columns    :List of 11
#>   ..$ mpg : chr "numeric"
#>   ..$ cyl : chr "numeric"
#>   ..$ disp: chr "numeric"
#>   ..$ hp  : chr "numeric"
#>   ..$ drat: chr "numeric"
#>   ..$ wt  : chr "numeric"
#>   ..$ qsec: chr "numeric"
#>   ..$ vs  : chr "numeric"
#>   ..$ am  : chr "numeric"
#>   ..$ gear: chr "numeric"
#>   ..$ carb: chr "numeric"
#>  $ path       : chr "mtcars"
#>  $ name       : chr "mtcars"
#>  $ api_version: int 0